+++ Tests/LinuxMain.swift
@@ -1,7 +1,0 @@
-import XCTest
-
-import iperf_swiftTests
-
-var tests = [XCTestCaseEntry]()
-tests += iperf_swiftTests.allTests()
-XCTMain(tests)